datastream refactor Fixes #16
